PDF and HTML can be generated seperately
