+++ scripts/templates/pdf-template.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <p>Version: {{ version }}</p>
-    {{{ content }}}
+    {{> symbols}}
   </body>
 </html>
