+++ scripts/templates/webpage-template.html
@@ -9,7 +9,8 @@
   </head>
   <body>
     <section class="main-content">
-      <a href="/">Back to Home</a> {{{ content }}}
+      <a href="/">Back to Home</a>
+      {{> symbols}}
     </section>
   </body>
 </html>
